refactor(langgraph): flatten adapter module names and cover the key-free example
Drop the underscore-prefixed private modules in favour of the flat, plainly
named modules used across the repo (powercontext.client, dsh), expressing
privacy through __all__ instead of filenames:

  _client.py   -> client.py
  _settings.py -> settings.py
  _runtime.py  -> runtime.py
  _scope.py    -> merged into scope.py

Merging _scope.py into scope.py removes the confusing scope/_scope pair and
colocates the PowerContextScope dataclass with its resolver, mirroring dsh's
single scope module.

Also add an e2e smoke test that runs the shipped, key-free inspect_recall
example end-to-end so it cannot rot as the adapter evolves.
