+++ integrations/langgraph/src/powercontext_langgraph/__init__.py
@@ -16,10 +16,9 @@
 
 from __future__ import annotations
 
-from ._scope import MissingScopeError, resolve_scope_id
-from ._settings import PowerContextLangGraphSettings
 from .recall import PowerContextRecall
-from .scope import PowerContextScope
+from .scope import MissingScopeError, PowerContextScope, resolve_scope_id
+from .settings import PowerContextLangGraphSettings
 from .tools import powercontext_context, powercontext_remember, powercontext_search, powercontext_tools
 
 __all__ = [
